+++ Kawai.Domain/DTOs/InventoryReportDto.cs
@@ -7,12 +7,12 @@ public class InventoryReportDto : DataTableDto
     public string ProductCode { get; set; }
     public string ProductName { get; set; }
     public string LotNo { get; set; }
-    public double PreMonth { get; set; }
-    public double Receipt { get; set; }
-    public double Supply { get; set; }
-    public double LossReject { get; set; }
-    public double Current { get; set; }
-    public double Inventory { get; set; }
+    public decimal PreMonth { get; set; }
+    public decimal Receipt { get; set; }
+    public decimal Supply { get; set; }
+    public decimal LossReject { get; set; }
+    public decimal Current { get; set; }
+    public decimal Inventory { get; set; }
     public string Remarks { get; set; }
     public string LastUser { get; set; }
 }
